+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 - Fixed accessory type update
 - Fixed synonyms delete
 - Improve PubChem structure images fetch
-
+- Fixed redirection when adding ingredient if already exists with the same name
 
 ### Version 12.4
 - Add dilutant in API formulas